requirements.txt
boto3==1.24.74
chowkidar-strawberry==0.2.2
coverage==6.5.0
cryptography==38.0.1
django==4.1.2
django-ipware==4.0.2
django-storages==1.13.1
environs==9.5.0
gunicorn==20.1.0
kramlin==0.1.4
psycopg2==2.9.4
Faker==15.1.0
strawberry-graphql==0.133.5
PyJWT==2.5.0
pycryptodome==3.15.0
requests==2.28.1
